Migrate Styling Configs (#2127)
* Migrate Styling Configs

* Fix styling

---------
