+++ src/Traits/Configuration/FooterConfiguration.php
@@ -47,18 +47,4 @@ trait FooterConfiguration
 
         return $this;
     }
-
-    public function setFooterTrAttributes(Closure $callback): self
-    {
-        $this->footerTrAttributesCallback = $callback;
-
-        return $this;
-    }
-
-    public function setFooterTdAttributes(Closure $callback): self
-    {
-        $this->footerTdAttributesCallback = $callback;
-
-        return $this;
-    }
 }
