+++ src/Traits/Configuration/PaginationConfiguration.php
@@ -154,13 +154,6 @@ trait PaginationConfiguration
         return $this;
     }
 
-    public function setPerPageFieldAttributes(array $attributes = []): self
-    {
-        $this->perPageFieldAttributes = [...$this->perPageFieldAttributes, ...$attributes];
-
-        return $this;
-    }
-
     public function setShouldRetrieveTotalItemCountStatus(bool $status): self
     {
         $this->shouldRetrieveTotalItemCount = $status;
@@ -179,13 +172,6 @@ trait PaginationConfiguration
     public function setShouldRetrieveTotalItemCountDisabled(): self
     {
         $this->setShouldRetrieveTotalItemCountStatus(false);
-
-        return $this;
-    }
-
-    public function setPaginationWrapperAttributes(array $paginationWrapperAttributes): self
-    {
-        $this->paginationWrapperAttributes = array_merge(['class' => ''], $paginationWrapperAttributes);
 
         return $this;
     }
