+++ src/Traits/Configuration/ReorderingConfiguration.php
@@ -81,14 +81,4 @@ trait ReorderingConfiguration
 
         return $this;
     }
-
-    /**
-     * Used to set attributes for the <th> for Reorder Column
-     */
-    public function setReorderThAttributes(array $reorderThAttributes): self
-    {
-        $this->reorderThAttributes = [...$this->reorderThAttributes, ...$reorderThAttributes];
-
-        return $this;
-    }
 }
