+++ src/Traits/Configuration/SecondaryHeaderConfiguration.php
@@ -2,8 +2,6 @@
 
 namespace Rappasoft\LaravelLivewireTables\Traits\Configuration;
 
-use Closure;
-
 trait SecondaryHeaderConfiguration
 {
     public function setSecondaryHeaderStatus(bool $status): self
@@ -23,20 +21,6 @@ trait SecondaryHeaderConfiguration
     public function setSecondaryHeaderDisabled(): self
     {
         $this->setSecondaryHeaderStatus(false);
-
-        return $this;
-    }
-
-    public function setSecondaryHeaderTrAttributes(Closure $callback): self
-    {
-        $this->secondaryHeaderTrAttributesCallback = $callback;
-
-        return $this;
-    }
-
-    public function setSecondaryHeaderTdAttributes(Closure $callback): self
-    {
-        $this->secondaryHeaderTdAttributesCallback = $callback;
 
         return $this;
     }
