+++ src/Traits/Helpers/ActionsHelpers.php
@@ -21,12 +21,6 @@ trait ActionsHelpers
     }
 
     #[Computed]
-    public function getActionWrapperAttributes(): array
-    {
-        return [...['class' => '', 'default-styling' => true, 'default-colors' => true], ...$this->actionWrapperAttributes];
-    }
-
-    #[Computed]
     public function hasActions(): bool
     {
         if (! isset($this->validActions)) {
