+++ src/Traits/Helpers/CollapsingColumnHelpers.php
@@ -26,28 +26,6 @@ trait CollapsingColumnHelpers
         return $this->getCollapsingColumnsStatus() === false;
     }
 
-    /**
-     * Retrieves attributes for the Collapsed Column Collapse Button
-     *
-     * @return array<mixed>
-     */
-    #[Computed]
-    public function getCollapsingColumnButtonCollapseAttributes(): array
-    {
-        return [...['default-styling' => true, 'default-colors' => true], ...$this->collapsingColumnButtonCollapseAttributes];
-    }
-
-    /**
-     * Retrieves attributes for the Collapsed Column Expand Button
-     *
-     * @return array<mixed>
-     */
-    #[Computed]
-    public function getCollapsingColumnButtonExpandAttributes(): array
-    {
-        return [...['default-styling' => true, 'default-colors' => true], ...$this->collapsingColumnButtonExpandAttributes];
-    }
-
     #[Computed]
     public function showCollapsingColumnSections(): bool
     {
