+++ src/Traits/Helpers/FooterHelpers.php
@@ -40,22 +40,4 @@ trait FooterHelpers
     {
         return $this->getUseHeaderAsFooterStatus() === false;
     }
-
-    /**
-     * @param  mixed  $rows
-     * @return array<mixed>
-     */
-    public function getFooterTrAttributes($rows): array
-    {
-        return isset($this->footerTrAttributesCallback) ? call_user_func($this->footerTrAttributesCallback, $rows) : ['default' => true];
-    }
-
-    /**
-     * @param  mixed  $rows
-     * @return array<mixed>
-     */
-    public function getFooterTdAttributes(Column $column, $rows, int $index): array
-    {
-        return isset($this->footerTdAttributesCallback) ? call_user_func($this->footerTdAttributesCallback, $column, $rows, $index) : ['default' => true];
-    }
 }
