+++ src/Traits/Helpers/SecondaryHeaderHelpers.php
@@ -2,8 +2,6 @@
 
 namespace Rappasoft\LaravelLivewireTables\Traits\Helpers;
 
-use Rappasoft\LaravelLivewireTables\Views\Column;
-
 trait SecondaryHeaderHelpers
 {
     public function hasColumnsWithSecondaryHeader(): bool
@@ -24,23 +22,5 @@ trait SecondaryHeaderHelpers
     public function secondaryHeaderIsDisabled(): bool
     {
         return $this->getSecondaryHeaderStatus() === false;
-    }
-
-    /**
-     * @param  mixed  $rows
-     * @return array<mixed>
-     */
-    public function getSecondaryHeaderTrAttributes($rows): array
-    {
-        return isset($this->secondaryHeaderTrAttributesCallback) ? call_user_func($this->secondaryHeaderTrAttributesCallback, $rows) : ['default' => true];
-    }
-
-    /**
-     * @param  mixed  $rows
-     * @return array<mixed>
-     */
-    public function getSecondaryHeaderTdAttributes(Column $column, $rows, int $index): array
-    {
-        return isset($this->secondaryHeaderTdAttributesCallback) ? call_user_func($this->secondaryHeaderTdAttributesCallback, $column, $rows, $index) : ['default' => true];
     }
 }
