+++ src/Traits/Styling/Configuration/ActionsStylingConfiguration.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Rappasoft\LaravelLivewireTables\Traits\Styling\Configuration;
+
+trait ActionsStylingConfiguration
+{
+    public function setActionWrapperAttributes(array $actionWrapperAttributes): self
+    {
+        $this->actionWrapperAttributes = [...$this->actionWrapperAttributes, ...$actionWrapperAttributes];
+
+        return $this;
+    }
+}
