+++ src/Traits/WithActions.php
@@ -5,13 +5,13 @@ namespace Rappasoft\LaravelLivewireTables\Traits;
 use Illuminate\Support\Collection;
 use Rappasoft\LaravelLivewireTables\Traits\Configuration\ActionsConfiguration;
 use Rappasoft\LaravelLivewireTables\Traits\Helpers\ActionsHelpers;
+use Rappasoft\LaravelLivewireTables\Traits\Styling\HasActionsStyling;
 
 trait WithActions
 {
     use ActionsConfiguration,
-        ActionsHelpers;
-
-    protected array $actionWrapperAttributes = ['class' => '', 'default-styling' => true, 'default-colors' => true];
+        ActionsHelpers,
+        HasActionsStyling;
 
     protected bool $displayActionsInToolbar = false;
