+++ src/Traits/WithCollapsingColumns.php
@@ -4,15 +4,13 @@ namespace Rappasoft\LaravelLivewireTables\Traits;
 
 use Rappasoft\LaravelLivewireTables\Traits\Configuration\CollapsingColumnConfiguration;
 use Rappasoft\LaravelLivewireTables\Traits\Helpers\CollapsingColumnHelpers;
+use Rappasoft\LaravelLivewireTables\Traits\Styling\HasCollapsingColumnsStyling;
 
 trait WithCollapsingColumns
 {
-    use CollapsingColumnConfiguration;
-    use CollapsingColumnHelpers;
+    use CollapsingColumnConfiguration,
+        CollapsingColumnHelpers,
+        HasCollapsingColumnsStyling;
 
     protected bool $collapsingColumnsStatus = true;
-
-    protected array $collapsingColumnButtonCollapseAttributes = ['default-styling' => true, 'default-colors' => true];
-
-    protected array $collapsingColumnButtonExpandAttributes = ['default-styling' => true, 'default-colors' => true];
 }
