+++ src/Traits/WithFooter.php
@@ -5,21 +5,19 @@ namespace Rappasoft\LaravelLivewireTables\Traits;
 use Closure;
 use Rappasoft\LaravelLivewireTables\Traits\Configuration\FooterConfiguration;
 use Rappasoft\LaravelLivewireTables\Traits\Helpers\FooterHelpers;
+use Rappasoft\LaravelLivewireTables\Traits\Styling\HasFooterStyling;
 
 trait WithFooter
 {
     use FooterConfiguration,
-        FooterHelpers;
+        FooterHelpers,
+        HasFooterStyling;
 
     protected bool $footerStatus = true;
 
     protected bool $useHeaderAsFooterStatus = false;
 
     protected bool $columnsWithFooter = false;
-
-    protected ?Closure $footerTrAttributesCallback;
-
-    protected ?Closure $footerTdAttributesCallback;
 
     public function setupFooter(): void
     {
