+++ src/Traits/WithSecondaryHeader.php
@@ -2,22 +2,19 @@
 
 namespace Rappasoft\LaravelLivewireTables\Traits;
 
-use Closure;
 use Rappasoft\LaravelLivewireTables\Traits\Configuration\SecondaryHeaderConfiguration;
 use Rappasoft\LaravelLivewireTables\Traits\Helpers\SecondaryHeaderHelpers;
+use Rappasoft\LaravelLivewireTables\Traits\Styling\HasSecondaryHeaderStyling;
 
 trait WithSecondaryHeader
 {
     use SecondaryHeaderConfiguration,
-        SecondaryHeaderHelpers;
+        SecondaryHeaderHelpers,
+        HasSecondaryHeaderStyling;
 
     protected bool $secondaryHeaderStatus = true;
 
     protected bool $columnsWithSecondaryHeader = false;
-
-    protected ?Closure $secondaryHeaderTrAttributesCallback;
-
-    protected ?Closure $secondaryHeaderTdAttributesCallback;
 
     public function bootedWithSecondaryHeader(): void
     {
